backend/requirements.txt
#############################################
# Production Requirements (Backend - FastAPI)
# Consolidated and pinned for deployment
#############################################

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Config and validation (Pydantic v2)
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator>=2.1.0

# Database and migrations
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Azure SDKs
azure-core==1.33.0
azure-identity==1.15.0
azure-mgmt-core==1.6.0
azure-mgmt-network==29.0.0
azure-mgmt-resource==23.0.1
azure-mgmt-storage>=21.0.0
azure-mgmt-subscription==3.1.0
azure-storage-blob==12.19.0

# AI integration
openai==1.3.7
tiktoken==0.5.2

# Security and auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==45.0.6
bcrypt==3.2.0

# Async and background tasks
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
tenacity==8.2.3
cachetools==5.3.2


